hoist translator into service and add fr.json
