+++ app/assets/messages/fr.json
@@ -1,0 +1,6 @@
+{
+  "/app/pages/index.js#title": "web 2 uniquement",
+  "/app/pages/index.js#apology": "Veuillez nous excuser pour le désordre.",
+  "/app/pages/index.js#explaination": "Nous sommes en train de tout reconstruire, littéralement.",
+  "/app/pages/index.js#call-to-action": "Suivre"
+}
